Rename login route file to loginLogout
