+++ src/server/app.js
@@ -17,7 +17,7 @@ const users = require('./routes/users');
 const readings = require('./routes/readings');
 const meters = require('./routes/meters');
 const preferences = require('./routes/preferences');
-const login = require('./routes/login');
+const login = require('./routes/loginLogout');
 const verification = require('./routes/verification');
 const groups = require('./routes/groups');
 const version = require('./routes/version');
